PUT restock (Service): forbidden (403) before bad-request (400) before not-found (404) PUT restock (Test-Service): changed assert comparisons; GET (Service): removed permission for check; GET (Test-Service): changed assert comparisons; Model: removed lower case for data validation; Error Handlers: separate module for error handlers Test err: separate test module for err
