+++ .pylintrc
@@ -1,2 +1,2 @@
 [MESSAGES CONTROL]
-disable=E1101, C0413, W0603, W1202, R0401, C0103
+disable=E1101, C0413, W0603, W1202, R0401, C0103, W0707
